+++ sdk/src/logs/log_record_data.cc
@@ -6,14 +6,15 @@
 #include <memory>
 #include <string>
 #include <unordered_map>
+#include <utility>
 
 #include "opentelemetry/common/attribute_value.h"
 #include "opentelemetry/common/timestamp.h"
 #include "opentelemetry/logs/severity.h"
 #include "opentelemetry/nostd/string_view.h"
-#include "opentelemetry/nostd/variant.h"
+#include "opentelemetry/sdk/common/attribute_utils.h"
 #include "opentelemetry/sdk/instrumentationscope/instrumentation_scope.h"
-#include "opentelemetry/sdk/logs/read_write_log_record.h"
+#include "opentelemetry/sdk/logs/log_record_data.h"
 #include "opentelemetry/sdk/resource/resource.h"
 #include "opentelemetry/trace/span_id.h"
 #include "opentelemetry/trace/trace_flags.h"
@@ -26,59 +27,60 @@ namespace sdk
 namespace logs
 {
 
-ReadWriteLogRecord::ReadWriteLogRecord()
+LogRecordData::LogRecordData()
     : severity_(opentelemetry::logs::Severity::kInvalid),
       resource_(nullptr),
       instrumentation_scope_(nullptr),
-#if OPENTELEMETRY_ABI_VERSION_NO >= 2
-      body_(std::string()),
-#else
-      body_(nostd::string_view()),
-#endif
       observed_timestamp_(std::chrono::system_clock::now()),
       event_id_(0),
       event_name_("")
-{}
+{
 
-ReadWriteLogRecord::~ReadWriteLogRecord() {}
+#if OPENTELEMETRY_ABI_VERSION_NO >= 2
+  body_ = std::string();
+#else
+  body_.SetAttribute("", nostd::string_view());
+#endif
+}
 
-void ReadWriteLogRecord::SetTimestamp(opentelemetry::common::SystemTimestamp timestamp) noexcept
+LogRecordData::~LogRecordData() {}
+
+void LogRecordData::SetTimestamp(opentelemetry::common::SystemTimestamp timestamp) noexcept
 {
   timestamp_ = timestamp;
 }
 
-opentelemetry::common::SystemTimestamp ReadWriteLogRecord::GetTimestamp() const noexcept
+opentelemetry::common::SystemTimestamp LogRecordData::GetTimestamp() const noexcept
 {
   return timestamp_;
 }
 
-void ReadWriteLogRecord::SetObservedTimestamp(
-    opentelemetry::common::SystemTimestamp timestamp) noexcept
+void LogRecordData::SetObservedTimestamp(opentelemetry::common::SystemTimestamp timestamp) noexcept
 {
   observed_timestamp_ = timestamp;
 }
 
-opentelemetry::common::SystemTimestamp ReadWriteLogRecord::GetObservedTimestamp() const noexcept
+opentelemetry::common::SystemTimestamp LogRecordData::GetObservedTimestamp() const noexcept
 {
   return observed_timestamp_;
 }
 
-void ReadWriteLogRecord::SetSeverity(opentelemetry::logs::Severity severity) noexcept
+void LogRecordData::SetSeverity(opentelemetry::logs::Severity severity) noexcept
 {
   severity_ = severity;
 }
 
-opentelemetry::logs::Severity ReadWriteLogRecord::GetSeverity() const noexcept
+opentelemetry::logs::Severity LogRecordData::GetSeverity() const noexcept
 {
   return severity_;
 }
 
-void ReadWriteLogRecord::SetBody(const opentelemetry::common::AttributeValue &message) noexcept
+void LogRecordData::SetBody(const opentelemetry::common::AttributeValue &message) noexcept
 {
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
   body_ = nostd::visit(attribute_converter_, message);
 #else
-  body_ = message;
+  body_.SetAttribute("", message);
 #endif
 }
 
@@ -87,28 +89,32 @@ const common::OwnedAttributeValue &
 #else
 const opentelemetry::common::AttributeValue &
 #endif
-ReadWriteLogRecord::GetBody() const noexcept
+LogRecordData::GetBody() const noexcept
 {
+#if OPENTELEMETRY_ABI_VERSION_NO >= 2
   return body_;
+#else
+  return body_.GetAttributes().begin()->second;
+#endif
 }
 
-void ReadWriteLogRecord::SetEventId(int64_t id, nostd::string_view name) noexcept
+void LogRecordData::SetEventId(int64_t id, nostd::string_view name) noexcept
 {
   event_id_   = id;
   event_name_ = std::string{name};
 }
 
-int64_t ReadWriteLogRecord::GetEventId() const noexcept
+int64_t LogRecordData::GetEventId() const noexcept
 {
   return event_id_;
 }
 
-nostd::string_view ReadWriteLogRecord::GetEventName() const noexcept
+nostd::string_view LogRecordData::GetEventName() const noexcept
 {
   return nostd::string_view{event_name_};
 }
 
-void ReadWriteLogRecord::SetTraceId(const opentelemetry::trace::TraceId &trace_id) noexcept
+void LogRecordData::SetTraceId(const opentelemetry::trace::TraceId &trace_id) noexcept
 {
   if (!trace_state_)
   {
@@ -118,7 +124,7 @@ void ReadWriteLogRecord::SetTraceId(const opentelemetry::trace::TraceId &trace_i
   trace_state_->trace_id = trace_id;
 }
 
-const opentelemetry::trace::TraceId &ReadWriteLogRecord::GetTraceId() const noexcept
+const opentelemetry::trace::TraceId &LogRecordData::GetTraceId() const noexcept
 {
   if (trace_state_)
   {
@@ -129,7 +135,7 @@ const opentelemetry::trace::TraceId &ReadWriteLogRecord::GetTraceId() const noex
   return empty;
 }
 
-void ReadWriteLogRecord::SetSpanId(const opentelemetry::trace::SpanId &span_id) noexcept
+void LogRecordData::SetSpanId(const opentelemetry::trace::SpanId &span_id) noexcept
 {
   if (!trace_state_)
   {
@@ -139,7 +145,7 @@ void ReadWriteLogRecord::SetSpanId(const opentelemetry::trace::SpanId &span_id) 
   trace_state_->span_id = span_id;
 }
 
-const opentelemetry::trace::SpanId &ReadWriteLogRecord::GetSpanId() const noexcept
+const opentelemetry::trace::SpanId &LogRecordData::GetSpanId() const noexcept
 {
   if (trace_state_)
   {
@@ -150,7 +156,7 @@ const opentelemetry::trace::SpanId &ReadWriteLogRecord::GetSpanId() const noexce
   return empty;
 }
 
-void ReadWriteLogRecord::SetTraceFlags(const opentelemetry::trace::TraceFlags &trace_flags) noexcept
+void LogRecordData::SetTraceFlags(const opentelemetry::trace::TraceFlags &trace_flags) noexcept
 {
   if (!trace_state_)
   {
@@ -160,7 +166,7 @@ void ReadWriteLogRecord::SetTraceFlags(const opentelemetry::trace::TraceFlags &t
   trace_state_->trace_flags = trace_flags;
 }
 
-const opentelemetry::trace::TraceFlags &ReadWriteLogRecord::GetTraceFlags() const noexcept
+const opentelemetry::trace::TraceFlags &LogRecordData::GetTraceFlags() const noexcept
 {
   if (trace_state_)
   {
@@ -171,14 +177,10 @@ const opentelemetry::trace::TraceFlags &ReadWriteLogRecord::GetTraceFlags() cons
   return empty;
 }
 
-void ReadWriteLogRecord::SetAttribute(nostd::string_view key,
-                                      const opentelemetry::common::AttributeValue &value) noexcept
+void LogRecordData::SetAttribute(nostd::string_view key,
+                                 const opentelemetry::common::AttributeValue &value) noexcept
 {
-#if OPENTELEMETRY_ABI_VERSION_NO >= 2
   attributes_map_.SetAttribute(key, value);
-#else
-  attributes_map_[std::string(key)] = value;
-#endif
 }
 
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
@@ -186,16 +188,12 @@ const std::unordered_map<std::string, opentelemetry::sdk::common::OwnedAttribute
 #else
 const std::unordered_map<std::string, opentelemetry::common::AttributeValue> &
 #endif
-ReadWriteLogRecord::GetAttributes() const noexcept
+LogRecordData::GetAttributes() const noexcept
 {
-#if OPENTELEMETRY_ABI_VERSION_NO >= 2
   return attributes_map_.GetAttributes();
-#else
-  return attributes_map_;
-#endif
 }
 
-const opentelemetry::sdk::resource::Resource &ReadWriteLogRecord::GetResource() const noexcept
+const opentelemetry::sdk::resource::Resource &LogRecordData::GetResource() const noexcept
 {
   if OPENTELEMETRY_LIKELY_CONDITION (nullptr != resource_)
   {
@@ -205,14 +203,13 @@ const opentelemetry::sdk::resource::Resource &ReadWriteLogRecord::GetResource() 
   return GetDefaultResource();
 }
 
-void ReadWriteLogRecord::SetResource(
-    const opentelemetry::sdk::resource::Resource &resource) noexcept
+void LogRecordData::SetResource(const opentelemetry::sdk::resource::Resource &resource) noexcept
 {
   resource_ = &resource;
 }
 
 const opentelemetry::sdk::instrumentationscope::InstrumentationScope &
-ReadWriteLogRecord::GetInstrumentationScope() const noexcept
+LogRecordData::GetInstrumentationScope() const noexcept
 {
   if OPENTELEMETRY_LIKELY_CONDITION (nullptr != instrumentation_scope_)
   {
@@ -222,7 +219,7 @@ ReadWriteLogRecord::GetInstrumentationScope() const noexcept
   return GetDefaultInstrumentationScope();
 }
 
-void ReadWriteLogRecord::SetInstrumentationScope(
+void LogRecordData::SetInstrumentationScope(
     const opentelemetry::sdk::instrumentationscope::InstrumentationScope
         &instrumentation_scope) noexcept
 {