+++ CHANGELOG.md
@@ -55,6 +55,22 @@ Increment the:
 * [CMAKE] Add generated protobuf headers to the opentelemetry_proto target
   [#3400](https://github.com/open-telemetry/opentelemetry-cpp/pull/3400)
 
+* [SDK] Fix lifetime for sdk::ReadWriteLogRecord
+  [#3147](https://github.com/open-telemetry/opentelemetry-cpp/pull/3245)
+
+Important changes:
+
+* [SDK] Fix lifetime for sdk::ReadWriteLogRecord
+  [#3147](https://github.com/open-telemetry/opentelemetry-cpp/pull/3245)
+
+  * With ABI version 2, `opentelemetry::sdk::logs::ReadableLogRecord::GetAttributes()`
+    will returns a `std::unordered_map<std::string, opentelemetry::sdk::common::OwnedAttributeValue>`
+    instead of a `std::unordered_map<std::string, opentelemetry::common::AttributeValue>`
+    to keep the same type as the `opentelemetry::sdk::trace::SpanData`. And
+    `opentelemetry::sdk::logs::ReadableLogRecord::GetBody()` will returns a
+    `const opentelemetry::sdk::common::OwnedAttributeValue &` instead of a
+    `const opentelemetry::common::AttributeValue &`.
+
 ## [1.20 2025-04-01]
 
 * [BUILD] Update opentelemetry-proto version
@@ -198,6 +214,9 @@ Important changes:
     `Provider` classes, to benefit from this feature.
 
   * All the example code has been updated to reflect the new usage.
+
+* [SDK] Implement spec: MetricFilter
+  [#3235](https://github.com/open-telemetry/opentelemetry-cpp/pull/3235)
 
 ## [1.19 2025-01-22]